refactor(cli): move loadSchema function
