+++ packages/cli/src/utils/project/loadSchema.ts
